Moved retrieval folder to info module, modified relevant imports
