--- Oanda/modules/training/retrieval/__init__.py
@@ -1,6 +1,0 @@
-from os.path import dirname, basename, isfile, join
-import glob
-from .history import *
-from .preprocessing import *
-modules = glob.glob(join(dirname(__file__), "*.py"))
-__all__ = [ basename(f)[:-3] for f in modules if isfile(f) and not f.startswith('_')]
